menu

+++ src/main/java/com/baf/data/entities/Article.java
@@ -6,5 +6,5 @@ public class Article {
     private int id;
     private String libelle;
     private int qteStock;
-    private int prix;
+    private double prix;
 }
